add sync test for a new observer syncing an existing network
